解决推荐歌单歌曲级别字段类型问题

+++ app/src/main/java/com/tobery/personalmusic/entity/home/RecommendListEntity.java
@@ -237,7 +237,7 @@ public class RecommendListEntity {
                 private int br;
                 private int fid;
                 private int size;
-                private int vd;
+                private double vd;
                 private int sr;
             }
 
@@ -247,7 +247,7 @@ public class RecommendListEntity {
                 private int br;
                 private int fid;
                 private int size;
-                private int vd;
+                private double vd;
                 private int sr;
             }
 
@@ -257,7 +257,7 @@ public class RecommendListEntity {
                 private int br;
                 private int fid;
                 private int size;
-                private int vd;
+                private double vd;
                 private int sr;
             }
